% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw_gwglm_cv.R
\name{bw_gwglm_cv}
\alias{bw_gwglm_cv}
\title{Bandwidth Selection via CV for Geographically Weighted Generalized Linear Model (GWGLM)}
\usage{
bw_gwglm_cv(
  formula,
  family,
  cordxy,
  distmethod = "euclidean",
  kernel,
  adaptive = TRUE,
  data,
  bw_min = NA,
  bw_max = NA,
  parallel = TRUE,
  core = NULL
)
}
\arguments{
\item{formula}{A formula object specifying the model structure (response ~ predictors).}

\item{family}{A character string specifying the distribution family. Supported options are \code{"gaussian"}, \code{"poisson"}, and \code{"binomial"}.}

\item{cordxy}{A matrix of spatial coordinates, with columns representing X and Y.}

\item{distmethod}{A character string indicating the distance calculation method to use.
Supported options are \code{"euclidean"} (default), \code{"greatcircle"} (great-circle distance), and \code{"manhattan"} (city-block distance).}

\item{kernel}{A character string specifying the kernel function used to compute spatial weights.
Options are \code{"global"}, \code{"gaussian"}, \code{"exponential"}, \code{"bisquare"}, \code{"tricube"}, and \code{"boxcar"}.}

\item{adaptive}{Logical. If \code{TRUE} (default), uses adaptive bandwidth. If \code{FALSE}, uses fixed bandwidth.}

\item{data}{A data frame containing all variables used in the model.}

\item{bw_min}{Minimum bandwidth value for the cross-validation search.
If \code{NA} (default), it is automatically set to the 10th percentile of non-zero distances (for fixed bandwidth) or 10\% of the number of observations (for adaptive bandwidth).}

\item{bw_max}{Maximum bandwidth value for the cross-validation search.
If \code{NA} (default), it is automatically set to the 90th percentile of non-zero distances (for fixed bandwidth) or 90\% of the number of observations (for adaptive bandwidth).}

\item{parallel}{Logical. If \code{TRUE} (default), enables parallel computation during the cross-validation procedure to improve speed.}

\item{core}{Integer. The number of CPU cores to use when \code{parallel = TRUE}.
If \code{NULL}, the number of cores is set to the total number of detected cores minus 2.}
}
\value{
A numeric value representing the optimal bandwidth.
}
\description{
This function selects the optimal bandwidth for a Geographically Weighted Generalized Linear Model (GWGLM)
by minimizing the cross-validation (CV) score using a golden section search algorithm.
It supports Gaussian, Poisson, and Binomial families, and offers both fixed and adaptive bandwidth options.
Parallel computing is supported to accelerate the CV evaluation process.
}
\details{
In addition to standard binary response variables of numeric \code{0/1} type,
this function and the associated GWGLM series uniquely support proportion-type responses
expressed as success counts and total trials (e.g., \code{y/n}).

When success/trial matrices are provided, the function automatically computes the success proportions internally,
eliminating the need for manual preprocessing.

This feature broadens the flexibility of modeling binary and proportion data,
addressing a common limitation in existing geographically weighted modeling packages.
}
\examples{
\dontrun{
data(tokyo)

tokyo$lnoff = log(tokyo$Exp_2564)

cordxy <- cbind(tokyo$X, tokyo$Y)
formula <- Mort2564~Professl+OwnHome+Elderly+Unemply+offset(lnoff)
family <- "poisson"


bw_cv <- bw_gwglm_cv(
  formula = formula,
  family = family,
  cordxy = cordxy,
  kernel = "gaussian",
  adaptive = FALSE,
  data = tokyo,
  parallel = TRUE
)
bw_cv

# y/n
data(infant)
cordxy <- cbind(infant$Xcoord, infant$Ycoord)

formula <- cbind(AVEDEATH03, AVEBIR03 - AVEDEATH03) ~ AVELBW03 +
            BLACK + HISPANIC + OTHERS + GINI + STABILITY
family <- "binomial"

bw_cv_2 <- bw_gwglm_cv(
formula = formula,
family = family,
cordxy = cordxy,
kernel = "gaussian",
adaptive = TRUE,
data = infant,
parallel = TRUE
)
bw_cv_2
}
}
\references{
#' Nakaya, T., Fotheringham, A. S., Charlton, M., & Brunsdon, C. (2009).
\emph{Semiparametric geographically weighted generalised linear modelling in GWR 4.0.}
}
