% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2vsim.R
\name{c2vsim.avHead}
\alias{c2vsim.avHead}
\title{c2vsim.avHead Calculates Average head values for selected time period}
\usage{
c2vsim.avHead(HeadList, ids)
}
\arguments{
\item{HeadList}{This is list of head. It is the third element of the output list from the
c2vsim.readHeadALL function}

\item{ids}{the ids of the months to consider in the averaging calculations}
}
\value{
A matrix [nNodes x nLay] with the average Head values
}
\description{
c2vsim.avHead Calculates Average head values for selected time period
}
