% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2vsim.R
\name{c2vsim.cumGWBUD}
\alias{c2vsim.cumGWBUD}
\title{c2vsim.cumGWBUD summarizes the groundwater budget for the entire domain or the
specified subregion ids}
\usage{
c2vsim.cumGWBUD(GWB, ids = NA)
}
\arguments{
\item{GWB}{This is a list of data frames with the budget terms for each
subregion. This is the output of the c2vsim.readGWBUD
function}

\item{ids}{the ids of the subregion you want to summarize}
}
\value{
A data frame with the budget terms corresponding to the entire area
}
\description{
c2vsim.cumGWBUD summarizes the groundwater budget for the entire domain or the
specified subregion ids
}
\examples{
GWB <- c2vsim.readGWBUD("CVground.BUD", Nsub = 21, Nskip = c(8,rep(9,20)), NtimeSteps = 1056, CG = TRUE)
GWBUDALL <- c2vsim.cumBUD(GWB, ids = c(5,10,21))
}
