% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2vsim.R
\name{c2vsim.cumLWBUD}
\alias{c2vsim.cumLWBUD}
\title{c2vsim.cumLWBUD calculates the cumulative lant and water use budget for the
selected subregion ids}
\usage{
c2vsim.cumLWBUD(LWB, ids = NA)
}
\arguments{
\item{LWB}{This is a list of data frames with the budget terms for each
subregion. This is the output of the c2vsim.readLWBUD
function}

\item{ids}{the ids of the subregion you want to cumulative budget.
Leave emtpty for the entire area}
}
\value{
A structure with the budget terms as they described in the c2vsim.readLWBUD
}
\description{
c2vsim.cumLWBUD calculates the cumulative lant and water use budget for the
selected subregion ids
}
\examples{
LWB <- c2vsim.readLWBUD("CVlandwater.BUD")
LW_ALL <- c2vsim.cumLWBUD(LWB)
}
