% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2vsim.R
\name{c2vsim.readDivData}
\alias{c2vsim.readDivData}
\title{c2vsim.readDivData Reads the Diversion Data file. The default values
correspond to the Coars version of C2Vsim input file}
\usage{
c2vsim.readDivData(filename, skiplines = 376, NtimeSteps = 1056)
}
\arguments{
\item{filename}{The name of the file}

\item{skiplines}{the number of files to skip up to the data}

\item{NtimeSteps}{the number of steps of the diversions timeseries}
}
\value{
a table of size NtimeSteps x the number of time series in file
}
\description{
c2vsim.readDivData Reads the Diversion Data file. The default values
correspond to the Coars version of C2Vsim input file
}
\examples{
divData <- c2vsim.readDivData("CVdiversions.dat")
}
