% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2vsim.R
\name{c2vsim.readDiversionBUD}
\alias{c2vsim.readDiversionBUD}
\title{c2vsim.readDiversionBUD}
\usage{
c2vsim.readDiversionBUD(filename, Nsub = 21, NtimeSteps = 1056, maxChar = 2000)
}
\arguments{
\item{filename}{is the name of the file.}

\item{Nsub}{the number of subregions printed in the file}

\item{NtimeSteps}{is the number of time steps used in the simulation. By
default is set to 1056}

\item{maxChar}{is the maximum number of characters that should read in
a single line}
}
\value{
A list of Nsub lists with the following fields:

Time: The time steps

SWD: Surface water deliveries.

DIV: Diversions

The above fields are list that containt part of the following fields

DivId: The diversion number

RivId: The river node id where the water is sxtracted from

SWD or DIV: the actual data for the surface water deliveries or diversions

SHRT: Shortage in the deliveries or diversions
}
\description{
Reads the Surface water deliveries and
diversion budget file.
}
\examples{
lw <- c2vsim.readDiversionBUD("CVdiverdtl.BUD")
}
