% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2vsim.R
\name{c2vsim.readDivSpec}
\alias{c2vsim.readDivSpec}
\title{c2vsim.readDivSpec reads the Diversion specification file.}
\usage{
c2vsim.readDivSpec(filename)
}
\arguments{
\item{filename}{is the name of the file to read}
}
\value{
a List with the following Fiels:
RDVnames: The names of the diversions
headers: a vector with the values NRDV, NDIVS, FACTX, FACTY
RDV: a matrix NRDV x 14 with the Surface Water Diversion Specifications
RDVELEM: a List of length NRDV with matrices Nelem x 2 that describe
         the Recharge zone for each diversion point
BYPS -> a matrix NDIVS x 6 with the Bypass Configuration Specifications
BYPSRT -> a list of NDIVS matrices with the rating tables for those bypasses that are defined.
BYPSELEM -> a list of NDIVS matrices Nelem x 2 that describe the Seepage
  locations for bypass canals
}
\description{
The way this function work is to read the entire file into the memory and
then process it line by line. However this maybe problematic for large files
see
(https://stackoverflow.com/questions/12626637/read-a-text-file-in-r-line-by-line)
}
\examples{
divSpec <- c2vsim.readDivSpec("CVdivspec.dat")
}
