% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2vsim.R
\name{c2vsim.readGWBUD}
\alias{c2vsim.readGWBUD}
\title{c2vsim.readGWBUD Reads the groundwater ascii budget file}
\usage{
c2vsim.readGWBUD(filename, Nsub = 21, Nskip = 8, NtimeSteps = 1056, CG = TRUE)
}
\arguments{
\item{filename}{is the filename}

\item{Nsub}{the number of subregions printed in the file}

\item{Nskip}{is the number of lines between the subregion tables.
In both versions the number of skip lines at the start of the file
is different than the number of skip lines between subregions. Luckily
This number is the same for all subregions.}

\item{NtimeSteps}{is the number of time steps used in the simulation. By
default is set to 1056}

\item{CG}{set this to true when reading from coarse grid version.
set it False when reading from fine grid}
}
\value{
Returns a list of size Nsub data frames with the budget time series.
  The columns in the data frame are the following:
  DP: Deep Percolation
  BS: Beginning Storage
  ES: Ending Storage
  NDP: Net Deep Percolation
  GFS: Gain From Stream
  R: Recharge
  GFL: Gain From Lake
  BI: Boundary Inflow
  S: Subsidence
  SI: Subsurface Irrigation
  TDO: Tile Drain Outflow
  P: Pumping
  NSI: Net Subsurface Inflow
  D: Discrepancy
  CS: Cumulative Subsidence
}
\description{
This file reads the groundwater budget file for the IWFM 302 version and
might not work for any model. However we have that works with the second beta
version of fine grid if you set the CG to false
}
\examples{
To read a coarse grid budget file we can use
c2vsim.readGWBUD("CVground.BUD", Nsub = 21, Nskip = c(8,rep(9,20)), NtimeSteps = 1056, CG = TRUE)
To read the fine grid budget file we modified as
c2vsim.readGWBUD("C2VSimFG_GW_Budget.bud", Nsub = 21, Nskip = c(8,rep(10,20)), NtimeSteps = 504, CG = FALSE)
}
