% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2vsim.R
\name{c2vsim.readGWHYD}
\alias{c2vsim.readGWHYD}
\title{c2vsim.readGWHYD Reads the groundwater hydrograph output file}
\usage{
c2vsim.readGWHYD(filename, Nskip = 4, NtimeSteps = 1056, maxChar = 40000)
}
\arguments{
\item{filename}{is the name of the groundwater hydrograph file.
For the coarse grid this is by default CVGWhyd.out}

\item{Nskip}{is the number of lines to skip before start reading the layer line}

\item{NtimeSteps}{is the number of time steps used in the simulation. By
default is set to 1056}

\item{maxChar}{is the maximum number of characters that should read in
a single line}
}
\value{
A list with the following fields:
Time: is the simulation times
LayerIds: is the layer of the groundwater node
NodeIds: is the groundwater node id
GWHyd: is a matrix NtimeSteps x m, where is m = length(LayerIds).
}
\description{
c2vsim.readGWHYD Reads the groundwater hydrograph output file
}
\examples{
For the coarse grid model we can read the hydrograph by using the default values:
GWHYD <- c2vsim.readGWHYD("CVGWhyd.out")
Lets suppose that we want to print the hydrograph for the groundwater node with id 536 for the first layer
First we have to find the column that corresponds to that node
i <- which(GWHYD$LayerIds == 1 & GWHYD$NodeIds == 536)
And then we can print as
plot(GWHYD$GWHyd[,i])
}
