% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2vsim.R
\name{c2vsim.readLWBUD}
\alias{c2vsim.readLWBUD}
\title{c2vsim.readLWBUD Reads the Land and water use budget output file for the coarse grid file.}
\usage{
c2vsim.readLWBUD(filename, Nsub = 21, NtimeSteps = 1056, maxchar = 2000)
}
\arguments{
\item{filename}{is the name of the file.}

\item{Nsub}{the number of subregions printed in the file}

\item{NtimeSteps}{is the number of time steps used in the simulation. By
default is set to 1056}

\item{maxChar}{is the maximum number of characters that should read in
a single line}
}
\value{
A list of Nsub lists with the following fields:
AG: a data frame with budget terms for the agricultural area
UR: a data frame with budget terms for the urban area
IE: a data frame with budget terms for subregion import/export
The data frames contain the following fields:
Area: The area of agricultural or ubran land
PCUAW: Potential CUAW
ASR: Agricultural supply requirement
P: Pumping
D: Diversion
S: Shortage
RU: ReUse
USR: Urban supply requirement
Import: Import
Export: Export
}
\description{
c2vsim.readLWBUD Reads the Land and water use budget output file for the coarse grid file.
}
\examples{
lw <- c2vsim.readLWBUD("CVlandwater.BUD")
}
