% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2vsim.R
\name{c2vsim.readMesh}
\alias{c2vsim.readMesh}
\title{c2vsim.readMesh Reads the Mesh element information of the C2Vsim coarse and fine grid}
\usage{
c2vsim.readMesh(filename, NE = 1392, Nskip = 93, Ncols = 5)
}
\arguments{
\item{filename}{is the name of the mesh element file. Usually is
CVelement.dat}

\item{NE}{is the number of elements. The default value corresponds to the
coarse grid version}

\item{Nskip}{is the number of lines to skip before start reading the list of
nodes. The default value corresponds to the coarse grid version.}

\item{Ncols}{is the number of columns. For the coarse grid version
is 5 : 1 for element ID and 4 columns for the node IDs. For the fine grid
there is an extra column with the subregion id.}
}
\value{
a data frame with the following fields: ID, ND1...ND4. If the
  elements are triangles then the ND4 index is zero
}
\description{
c2vsim.readMesh Reads the Mesh element information of the C2Vsim coarse and fine grid
}
\examples{
MSH <- c2vsim.readMesh("CVelement.dat")
}
