% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2vsim.R
\name{c2vsim.readNodes}
\alias{c2vsim.readNodes}
\title{c2vsim.Nodes Reads the Node coordinates of the C2Vsim}
\usage{
c2vsim.readNodes(filename, ND = 1393, Nskip = 80)
}
\arguments{
\item{filename}{is the name of the file. Usually is CVnode.dat}

\item{ND}{is the number of nodes. The default value corresponds to the coarse
grid version}

\item{Nskip}{is the number of lines to skip before start reading the list of
nodes, The default value corresponds to the coarse grid version}
}
\value{
a data frame with the following fields: ID, X and Y
}
\description{
c2vsim.Nodes Reads the Node coordinates of the C2Vsim
}
\examples{
XY <- c2vsim.Nodes("CVnode.dat")
}
