% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2vsim.R
\name{c2vsim.readParam}
\alias{c2vsim.readParam}
\title{c2vsim.readParam Reads parts of the parameter file.}
\usage{
c2vsim.readParam(filename, Nnodes = 1393, Nlay = 3, nSkip = 267)
}
\arguments{
\item{filename}{is the name of the stratigraphy file. Usually is
CVstrat.dat}

\item{Nnodes}{is the number of nodes.}

\item{Nlay}{is the number or layers.}

\item{Nskip}{is the number of lines to skip before start reading the list of
parameters. For the coarse grid version 267 for reading the saturated parameters}
}
\value{
A List of Nlay dataframes with the following fields
PKH: Hydraulic conductivity
PS: Specific storage
PN: Specific yield
PV: Aquitard vertical hydraulic conductivity
PL: Aquifer vertical hydraulic conductivity
SCE: Elastic storage coefficient (Use SCE*DC if DC=0)
SCI: Inelastic storage coefficient (Use SCI*DC if DC=0)
DC: Interbed thickness
DCMIN: Minimum interbed thickness
HC: Pre-compaction hydraulic head
}
\description{
This function does NOT read the entire file. Instead it reads only
the parameters that are setup based on OPTION 2
}
\examples{
param <-  c2vsim.readParam("CVparam.dat")
}
