% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2vsim.R
\name{c2vsim.readStrat}
\alias{c2vsim.readStrat}
\title{c2vsim.readStrat Reads the stratigraphy file.}
\usage{
c2vsim.readStrat(filename, Nnodes = 1393, Nlay = 3, nSkip = 92)
}
\arguments{
\item{filename}{is the name of the stratigraphy file. Usually is
CVstrat.dat}

\item{Nnodes}{is the number of nodes.}

\item{Nlay}{is the number or layers. Each layer consist of two values.
One for the main aquifer and one for the aquiclude}

\item{Nskip}{is the number of lines to skip before start reading the list of
nodes. The default value corresponds to the coarse grid version.}
}
\value{
a data frame with the following fields: ID, L11, L12,...L31,L32,...
}
\description{
The default parameter values corresond to coarse grid version
}
\examples{
STRAT <-  c2vsim.readStrat("CVstrat.dat")
}
