% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2vsim.R
\name{c2vsim.readSWHYD}
\alias{c2vsim.readSWHYD}
\title{c2vsim.readSWHYD Reads the stream hydrograph output file}
\usage{
c2vsim.readSWHYD(filename, Nskip = 5, NtimeSteps = 1056, maxChar = 7000)
}
\arguments{
\item{filename}{is the name of the stream hydrograph file.
For the coarse grid this is by default CVSWhyd.out}

\item{Nskip}{is the number of lines to skip before start reading the line with the stream node ids}

\item{NtimeSteps}{is the number of time steps used in the simulation. By
default is set to 1056}

\item{maxChar}{is the maximum number of characters that should read in
a single line}
}
\value{
A list with the following fields:
Time: The time steps
NodeIds: is a vector with the stream node ids
SWHyd: is a matrix NtimeSteps x m, where is m = length(NodeIds).
}
\description{
c2vsim.readSWHYD Reads the stream hydrograph output file
}
\examples{
For the coarse grid model we can read the stream hydrograph by using the default values:
SWHYD <- c2vsim.readSWHYD("CVSWhyd.out")
}
