% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2vsim.R
\name{c2vsim.writeDivData}
\alias{c2vsim.writeDivData}
\title{c2vsim.writeDivData WriteS the Diversion Data file.}
\usage{
c2vsim.writeDivData(
  filename,
  data,
  NCOLDV = 265,
  FACTDV = 43560000,
  NSPDV = 1,
  NFQDV = 0,
  DSSFL = "",
  sep = " "
)
}
\arguments{
\item{filename}{The name of the file}

\item{data}{Is a table similar to the output of c2vsim.readDivDatafunction.
Note that the NCOLDV and the size the data must be in agreement}

\item{NCOLDV}{FACTDV NSPDV NFQDV DSSFL : these are parameters of the input file.
See a template file for detail explanation of these.}

\item{NtimeSteps}{the number of steps of the diversions timeseries}
}
\value{
nothing
}
\description{
c2vsim.writeDivData WriteS the Diversion Data file.
}
