% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2vsim.R
\name{c2vsim.writeDivSpec}
\alias{c2vsim.writeDivSpec}
\title{c2vsim.writeDivSpec WriteS the Diversion Data file.}
\usage{
c2vsim.writeDivSpec(filename, DivSpec)
}
\arguments{
\item{filename}{The name of the file}

\item{DivSpec}{Is a structure similar to the output of the c2vsim.readDivSpec function.}
}
\value{
nothing
}
\description{
c2vsim.writeDivSpec WriteS the Diversion Data file.
}
