% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{distPointLineSeg}
\alias{distPointLineSeg}
\title{distPointLineSeg calculates the distance between a point and a line segment
defined by two points}
\usage{
distPointLineSeg(xp, yp, x1, y1, x2, y2)
}
\arguments{
\item{xp}{the x coordinate of the point}

\item{yp}{the y coordinate of the point}

\item{x1}{the x coordinate of the end point of the line}

\item{y1}{the y coordinate of the end point of the line}

\item{x2}{the x coordinate of the other end point of the line}

\item{y2}{the y coordinate of the other end point of the line}
}
\value{
the closest distance between the point and the line segment
}
\description{
The function projects the point onto the line and returns the perpendicular distance
between the point and the line if the projection fall within the line segment.
If the projection falls outside the line segments it returns the closest distance between
the point and the closest end segment point
}
