% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwglm.R
\name{gwglm}
\alias{gwglm}
\title{Geographically Weighted Generalized Linear Model (GWGLM)}
\usage{
gwglm(
  formula,
  family,
  cordxy,
  distmethod = "euclidean",
  h,
  kernel,
  adaptive = TRUE,
  cvar = 0,
  data,
  parallel = TRUE,
  core = NULL
)
}
\arguments{
\item{formula}{A formula object specifying the model structure (response ~ predictors).}

\item{family}{A character string specifying the distribution family. Supported options are \code{"gaussian"}, \code{"poisson"}, and \code{"binomial"}.}

\item{cordxy}{A matrix of spatial coordinates, with columns representing X and Y.}

\item{distmethod}{A character string indicating the distance calculation method to use.
Supported options are \code{"euclidean"} (default), \code{"greatcircle"} (great-circle distance), and \code{"manhattan"} (city-block distance).}

\item{h}{A numeric value specifying the bandwidth. This can represent a fixed distance or a number of nearest neighbors (if \code{adaptive = TRUE}).}

\item{kernel}{A character string specifying the kernel function used to compute spatial weights.
Options are \code{"global"}, \code{"gaussian"}, \code{"exponential"}, \code{"bisquare"}, \code{"tricube"}, and \code{"boxcar"}.}

\item{adaptive}{Logical. If \code{TRUE} (default), uses adaptive bandwidth. If \code{FALSE}, uses fixed bandwidth.}

\item{cvar}{A numeric value. An optional adjustment to the effective number of parameters,
commonly used when some variables are modeled with fixed (non-spatially varying) coefficients. The default is \code{0}.}

\item{data}{A data frame containing all variables used in the model.}

\item{parallel}{Logical. If \code{TRUE} (default), enables parallel computation to speed up local estimation. Default is \code{TRUE}.}

\item{core}{Integer. The number of CPU cores to use when \code{parallel = TRUE}.
If \code{NULL}, the number of cores is set to the total number of detected cores minus 2.}
}
\value{
An object of class \code{"gwglm_model"}, which is a list containing:
\itemize{
\item \code{formula}: The model formula used.
\item \code{beta}: Matrix of locally estimated regression coefficients.
\item \code{std}: Matrix of standard errors.
\item \code{tvalue}: Matrix of local t-values.
\item \code{pvalue}: Matrix of local p-values.
\item \code{muhat}: Vector of fitted values.
\item \code{error}: Vector of residuals.
\item \code{deviance}: Model deviance.
\item \code{logLik}: Model log-likelihood.
\item \code{AIC}: Akaike Information Criterion.
\item \code{AICc}: Corrected Akaike Information Criterion.
\item \code{BIC}: Bayesian Information Criterion.
\item \code{k}: Effective degrees of freedom (EDF).
\item \code{call}: The matched call that created the model.
}
}
\description{
This function fits a Geographically Weighted Generalized Linear Model (GWGLM),
where regression coefficients are locally estimated at each spatial location using kernel-weighted likelihood.
It supports Gaussian, Poisson, and Binomial families and returns local coefficient estimates along with standard errors,
t-values, p-values, fitted values, and model diagnostics such as AIC, AICc, BIC, and effective degrees of freedom (EDF).
}
\details{
In addition to standard binary response variables of numeric \code{0/1} type,
this function and the associated GWGLM series uniquely support proportion-type responses
expressed as success counts and total trials (e.g., \code{y/n}).

When success/trial matrices are provided, the function automatically computes the success proportions internally,
eliminating the need for manual preprocessing.

This feature broadens the flexibility of modeling binary and proportion data,
addressing a common limitation in existing geographically weighted modeling packages.
}
\examples{
\dontrun{
data(tokyo)

tokyo$lnoff = log(tokyo$Exp_2564)

cordxy <- cbind(tokyo$X, tokyo$Y)
formula <- Mort2564~Professl+OwnHome+Elderly+Unemply+offset(lnoff)
family <- "poisson"


bw_cv <- bw_gwglm_cv(
  formula = formula,
  family = family,
  cordxy = cordxy,
  kernel = "gaussian",
  adaptive = FALSE,
  data = tokyo,
  parallel = TRUE
)

gwglm_result <- gwglm(
  formula = formula,
  family = family,
  cordxy = cordxy,
  kernel = "gaussian",
  adaptive= FALSE,
  h = bw_cv,
  data = tokyo,
  parallel = T,
  core = 4
)

gwglm_result

# y/n
data(infant)
cordxy <- cbind(infant$Xcoord, infant$Ycoord)

formula <- cbind(AVEDEATH03, AVEBIR03 - AVEDEATH03) ~ AVELBW03 +
            BLACK + HISPANIC + OTHERS + GINI + STABILITY
family <- "binomial"

bw_cv_2 <- bw_gwglm_cv(
formula = formula,
family = family,
cordxy = cordxy,
kernel = "gaussian",
adaptive = TRUE,
data = infant,
parallel = TRUE
)
bw_cv_2

gwglm_result_2 <- gwglm(
  formula = formula,
  family = family,
  cordxy = cordxy,
  kernel = "gaussian",
  adaptive= FALSE,
  h = bw_cv,
  data = infant,
  parallel = T,
  core = 4
)

gwglm_result_2
}
}
\references{
Nakaya, T., Fotheringham, A. S., Charlton, M., & Brunsdon, C. (2009).
\emph{Semiparametric geographically weighted generalised linear modelling in GWR 4.0.}

Zheng, G.-T. (2024). \emph{An application of semi-parametric geographically weighted logistic regression in real estate transaction data analysis}.
Master's thesis, Department of Statistics, National Chengchi University, Taiwan.
}
