% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ind2sub}
\alias{ind2sub}
\title{Convert linear indices to subscripts in column major order}
\usage{
t <- ind2sub(j,n), then A[t$r, t$c] == A[l] if dim(A)[1] == n
}
\arguments{
\item{l}{linear index}

\item{n}{number of rows}
}
\value{
the linear index.
}
\description{
This function is somewhat the equivalent of the Matlab ind2sub.
}
