% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantis.R
\name{mantis.Ping}
\alias{mantis.Ping}
\title{mantis.Ping Test the communication with server.
It sends a ping and expects a pong}
\usage{
mantis.Ping(scenario)
}
\arguments{
\item{scenario}{the only useful option for this function
is the client which is is the path to the client executable}
}
\value{
Returns a pong if the communication is succesful
}
\description{
mantis.Ping Test the communication with server.
It sends a ping and expects a pong
}
\examples{
scenario <- mantis.ScenarioOptions()
scenario$client <- 'MantisClient.exe'
mantis.Ping <- function( scenario)
}
