% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantis.R
\name{mantis.Quit}
\alias{mantis.Quit}
\title{mantis.Quit Stops the server}
\usage{
mantis.Quit(client_prog)
}
\arguments{
\item{client_prog}{is the path to the Mantis client executable}
}
\value{
Returns nothing. It just quits the server
}
\description{
mantis.Quit Stops the server
}
\examples{
mantis.Quit <- function( 'MantisClient.exe')
}
