% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantis.R
\name{mantis.ReadOutput}
\alias{mantis.ReadOutput}
\title{mantis.ReadOutput reads the output file of the mantis client.
This is called by the mantis.Run and there is no reason why
someone would need to run this}
\usage{
mantis.ReadOutput(filename)
}
\arguments{
\item{filename}{This is the output file. The file name is specified
with the option scenario$outfile}
}
\value{
returns a table with the wells break though curves.
The number of rows is equal to the number of wells in the selected area
and the number of columns is the number of simulation years
}
\description{
mantis.ReadOutput reads the output file of the mantis client.
This is called by the mantis.Run and there is no reason why
someone would need to run this
}
\examples{
mantis.ReadOutput <- function(scenario$outfile)
}
