% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantis.R
\name{mantis.ReadWells}
\alias{mantis.ReadWells}
\title{mantis.ReadWells reads the well input file.}
\usage{
mantis.ReadWells(filename)
}
\arguments{
\item{filename}{This is the name of the input file.}
}
\value{
returns a table with the info for the wells.
The fields are:

ID: well id

X, Y: Well coordinate in EPSG 3310

Depth: Depth of the well. This is the distance between the simulated water table and the top of the screen

SL: Screen length. Distance between top and bottom of the well screen

Ratio, Angle: These two parameter define the shape of the well source area

The remaining parameters show in which sub area the well belongs to.
}
\description{
mantis.ReadWells reads the well input file.
}
\examples{
mantis.ReadWells <- function(C2VSIM_II_02_VI_Wells.mnts)
}
