% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantis.R
\name{mantis.Run}
\alias{mantis.Run}
\title{mantis.Run Runs the scenario and returns the results}
\usage{
mantis.Run(scenario, client_prog)
}
\arguments{
\item{scenario}{is a list with the scenario options}

\item{client_prog}{is the path to the Mantis client executable}
}
\value{
a list with the well breakthrough curves
}
\description{
mantis.Run Runs the scenario and returns the results
}
\examples{
result <- mantis.Run(scenario, 'MantisClient.exe')
}
