% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantis.R
\name{mantis.ScenarioOptions}
\alias{mantis.ScenarioOptions}
\title{mantis.ScenarioOptions Returns a structure with all available Mantis Scenario options}
\usage{
mantis.ScenarioOptions()
}
\value{
A data frame with the budget terms corresponding to the entire area
}
\description{
mantis.ScenarioOptions Returns a structure with all available Mantis Scenario options
}
\examples{
GWB <- c2vsim.readGWBUD("CVground.BUD", Nsub = 21, Nskip = c(8,rep(9,20)), NtimeSteps = 1056, CG = TRUE)
GWBUDALL <- c2vsim.cumBUD(GWB, ids = c(5,10,21))
}
