% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantis.R
\name{mantis.WriteInput}
\alias{mantis.WriteInput}
\title{mantis.WriteInput writes the input file for a given scenario.
This is called by the mantis.Run and there is no reason why
someone would need to run this}
\usage{
mantis.WriteInput(scenario)
}
\arguments{
\item{scenario}{is a list with the scenario options}

\item{client_prog}{is the path to the Mantis client executable}
}
\value{
It returns nothing and prints the input file.
The name of the input file is set by the option scenario$infile
}
\description{
mantis.WriteInput writes the input file for a given scenario.
This is called by the mantis.Run and there is no reason why
someone would need to run this
}
\examples{
mantis.WriteInput <-function(scenario)
}
