% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modflow.R
\name{modflow.readArrayASCII}
\alias{modflow.readArrayASCII}
\title{modflow.readArrayASCII reads a 3D array that is printed in ASCII.
This can be used to read the modflow heads when they are printed as ASCII}
\usage{
modflow.readArrayASCII(filename, nlay, maxchar = 1000)
}
\arguments{
\item{filename}{is the file name}

\item{nlay}{The number of layers}

\item{maxchar}{A guess about the maximum width of the lines}
}
\value{
A list with one 3D array per time step
}
\description{
modflow.readArrayASCII reads a 3D array that is printed in ASCII.
This can be used to read the modflow heads when they are printed as ASCII
}
\examples{
gwHead <- modflow.readArrayASCII(headsout.txt, nlay = 13)
}
