% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modflow.R
\name{modflow.readCBC}
\alias{modflow.readCBC}
\title{modflow.readCBC reads the cell by cell modflow binary file
See more on https://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?frequently_asked_questions.htm
expand L and then expand Flow Data}
\usage{
modflow.readCBC(filename, endian = "little", intsize = 4)
}
\arguments{
\item{filename}{is the file name}

\item{endian}{should be either "little" or "big"}

\item{intsize}{number of bytes for the integer numbers}
}
\value{
A complicated List
}
\description{
modflow.readCBC reads the cell by cell modflow binary file
See more on https://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?frequently_asked_questions.htm
expand L and then expand Flow Data
}
\examples{
cbc <- modflow.readCB(cbc.out)
}
