% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlo_gwglm.R
\name{montecarlo_gwglm}
\alias{montecarlo_gwglm}
\title{Monte Carlo Test for Spatial Variablility in GWGLM}
\usage{
montecarlo_gwglm(
  formula,
  family,
  cordxy,
  distmethod = "euclidean",
  h,
  kernel,
  adaptive,
  data,
  n_permutations = 99,
  parallel = TRUE,
  core = NULL
)
}
\arguments{
\item{formula}{A formula object specifying the model structure (response ~ predictors).}

\item{family}{A character string specifying the distribution family. Supported options are \code{"gaussian"}, \code{"poisson"}, and \code{"binomial"}.}

\item{cordxy}{A matrix of spatial coordinates, with columns representing X and Y.}

\item{distmethod}{A character string indicating the distance calculation method to use.
Supported options are \code{"euclidean"} (default), \code{"greatcircle"} (great-circle distance), and \code{"manhattan"} (city-block distance).}

\item{h}{A numeric value specifying the bandwidth. This can represent a fixed distance or a number of nearest neighbors (if \code{adaptive = TRUE}).}

\item{kernel}{A character string specifying the kernel function used to compute spatial weights.
Options are \code{"global"}, \code{"gaussian"}, \code{"exponential"}, \code{"bisquare"}, \code{"tricube"}, and \code{"boxcar"}.}

\item{adaptive}{Logical. If \code{TRUE} (default), uses adaptive bandwidth. If \code{FALSE}, uses fixed bandwidth.}

\item{data}{A data frame containing all variables used in the model.}

\item{n_permutations}{Number of Monte Carlo permutations. Default is 99.}

\item{parallel}{Logical. If \code{TRUE} (default), enables parallel computation to speed up local estimation. Default is \code{TRUE}.}

\item{core}{Integer. The number of CPU cores to use when \code{parallel = TRUE}.
If \code{NULL}, the number of cores is set to the total number of detected cores minus 2.}
}
\value{
A list with one element:
\describe{
\item{\code{p_values}}{A named vector of p-values for each coefficient, indicating whether its spatial variation is statistically significant.}
}
}
\description{
This function performs a Monte Carlo test to examine whether the local coefficients
in a Geographically Weighted Generalized Linear Model (GWGLM) vary significantly across different locations.
}
\details{
The function first computes the observed variance of each local coefficient across space.
Then, it repeatedly permutes the spatial coordinates and refits the model to simulate a null distribution
of coefficient variances under spatial randomness.
For each coefficient, a \code{p}-value is calculated by comparing the observed variance
to the distribution of permuted variances.

In addition to standard binary response variables of numeric \code{0/1} type,
this function and the associated GWGLM series uniquely support proportion-type responses
expressed as success counts and total trials (e.g., \code{y/n}).

When success/trial matrices are provided, the function automatically computes the success proportions internally,
eliminating the need for manual preprocessing.

This feature broadens the flexibility of modeling binary and proportion data,
addressing a common limitation in existing geographically weighted modeling packages.
}
\examples{
\dontrun{
data(tokyo)

tokyo$lnoff = log(tokyo$Exp_2564)

cordxy <- cbind(tokyo$X, tokyo$Y)
formula <- Mort2564~Professl+OwnHome+Elderly+Unemply+offset(lnoff)
family <- "poisson"

bw_cv <- bw_gwglm_cv(
  formula = formula,
  family = family,
  cordxy = cordxy,
  kernel = "gaussian",
  adaptive = FALSE,
  data = tokyo,
  parallel = TRUE
)

montecarlo_result <- montecarlo_gwglm(
  formula = formula,
  family = family,
  cordxy = cordxy,
  h = bw_cv,
  kernel = "gaussian",
  adaptive= FALSE,
  data = tokyo,
  n_permutations = 99,
  parallel = T,
  core = 4
)
montecarlo_result
}
}
