% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsat.R
\name{npsat.ReadScattered}
\alias{npsat.ReadScattered}
\title{npsat.ReadScattered reads the scattered format files}
\usage{
npsat.ReadScattered(filename, PDIM = 2)
}
\arguments{
\item{filename}{is the name of the file}

\item{PDIM}{is the dimention of the point. Valid options are 2 for 2D points or 1 for 1D points.
At the moment the points can be only 1D or 2D.}
}
\value{
a data frame with the X Y Values
}
\description{
npsat.ReadScattered reads the scattered format files
}
