% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsat.R
\name{npsat.ReadStreams}
\alias{npsat.ReadStreams}
\title{npsat.ReadStreams reads the Stream file}
\usage{
npsat.ReadStreams(filename, maxChar = 1000)
}
\arguments{
\item{filename}{is the name of the file}

\item{maxChar}{is the maximum number of characters in a single line}
}
\value{
A list of Stream polygons with the following fields:
Coords: the coordinates of the stream polygon
Rate: The rate in L/T units
Area the area of the polygon
Q: The total volume of water in L^3/T
Info: This can be any info associated with the river segments
}
\description{
npsat.ReadStreams reads the Stream file
}
