% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsat.R
\name{npsat.ReadWaterTableCloudPoints}
\alias{npsat.ReadWaterTableCloudPoints}
\title{npsat.ReadWaterTableCloudPoints reads the water table cloud point file}
\usage{
npsat.ReadWaterTableCloudPoints(
  prefix,
  suffix = "top",
  nproc = 1,
  iter = 1,
  print_file = F
)
}
\arguments{
\item{prefix}{is the prefix used in the simulation input file}

\item{suffix}{This is 'top' by default}

\item{nproc}{The number of processors used in the simulation}

\item{iter}{the iteration number}

\item{print_file}{This is currently not used}
}
\value{
A dataframe with the following fields:
\itemize{
    \item X, Y: coordinates of the node
    \item Zold: the initial elevation
    \item Znew: the new elevation
}
}
\description{
npsat.ReadWaterTableCloudPoints reads the water table cloud point file
}
