% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsat.R
\name{npsat.ReadWells}
\alias{npsat.ReadWells}
\title{npsat.ReadWells reads the well file}
\usage{
npsat.ReadWells(filename)
}
\arguments{
\item{filename}{is the name of the file}

\item{maxChar}{is the maximum number of characters in a single line}
}
\value{
A dataframe with the wells with the following fields:
\itemize{
    \item X, Y: coordinates of the well
    \item T: top elevation of the well screen
    \item B: bottom elevation of the well screen
    \item Q: The pumping rate in L^3/T
}
}
\description{
npsat.ReadWells reads the well file
}
