% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsat.R
\name{npsat.WriteGridAxis}
\alias{npsat.WriteGridAxis}
\title{npsat.WriteGridAxis writes a xis object to file}
\usage{
npsat.WriteGridAxis(filename, x, spacing)
}
\arguments{
\item{filename}{is the file name}

\item{x}{If the spacing option is CONST then x is a length 3 vector
where the x[1] is the origin x[2] is the distance between ticks and x[3] is the number of ticks
in that axis. If spacing is VAR then x must be a monotonous increasing vector with the ticks.
This is a good option if the ticks are not evenly spaced}

\item{spacing}{spacing is either CONST or VAR}
}
\value{
just prints a file
}
\description{
npsat.WriteGridAxis writes a xis object to file
}
\examples{
npsat.WriteGridAxis("axis1.dat, c(0,100,23), "CONST")
will generate an axis with ticks starting from zero to 2200
npsat.WriteGridAxis("axis1.dat, x, "VAR")
will print the vector x as the axis. Note that it is not checked for doubles or
non monotonous increase in this function neither by the program itself
}
