% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsat.R
\name{npsat.writeMesh}
\alias{npsat.writeMesh}
\title{Writes a mesh into a file for reading from the npsat_engine}
\usage{
npsat.writeMesh(filename, nd, msh)
}
\arguments{
\item{filename}{is the name of the file to write the mesh.}

\item{nd}{A matrix n x 2 with the mesh node coordinates}

\item{msh}{A matrix with the indices of the mesh elements. The matrix must be based on 1.
However the file will be printed as zero based}
}
\value{
nothing
}
\description{
Writes a mesh into a file for reading from the npsat_engine
}
