% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsat.R
\name{npsat.WriteWells}
\alias{npsat.WriteWells}
\title{npsat.WriteWells writes the wells to a file}
\usage{
npsat.WriteWells(filename, wells)
}
\arguments{
\item{filename}{is the name of the file}

\item{wells}{is a data frame with the following files in the following order
X coordinate
Y coordinate
Top elevation of the top of the screen
Bottom elevation of the bottom of the screen
Q pumping rate. Extraction is negative}
}
\value{
nothing
}
\description{
npsat.WriteWells writes the wells to a file
}
