% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsgaii.R
\name{nsgaii.readParetoHistory}
\alias{nsgaii.readParetoHistory}
\title{nsgaii.readParetoHistory
Reads the pareto history file}
\usage{
nsgaii.readParetoHistory(filename, nGen)
}
\arguments{
\item{filename}{The file where the Pareto history is written}

\item{nGen}{Then total number of generations}
}
\value{
a list of lenght nGew with a matrix in each entry representing the
  current pareto front
}
\description{
nsgaii.readParetoHistory
Reads the pareto history file
}
