% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsgaii.R
\name{nsgaii.readParetoSolution}
\alias{nsgaii.readParetoSolution}
\title{nsgaii.readParetoSolution reads the pareto solution.}
\usage{
nsgaii.readParetoSolution(filename)
}
\arguments{
\item{filename}{The file name is the OutputFile parameter of the configuration file of nsgaii}
}
\value{
a list with 2 elements. dv contains the decision variables and of the
objective function values
}
\description{
nsgaii.readParetoSolution reads the pareto solution.
}
