% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bootstrap_TSGWML.R
\name{print.bootstrap_TSGWML}
\alias{print.bootstrap_TSGWML}
\title{Print Method for TSGWML Bootstrap Test Results}
\usage{
\method{print}{bootstrap_TSGWML}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"bootstrap_TSGWML"} returned by \code{\link{bootstrap_TSGWML}}.}

\item{...}{Further arguments passed to or from other methods. Currently ignored.}
}
\value{
Invisibly returns \code{NULL}. Called for its side effect of printing model output to the console.
}
\description{
Prints a brief summary of a \code{bootstrap_TSGWML} object,
including the function call, model formula, variable types, the observed likelihood ratio test (LRT) statistic,
the associated bootstrap p-value, and the bootstrap-based standard errors for both global and local coefficients.
}
\seealso{
\code{\link{bootstrap_TSGWML}}
}
