% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gwmmr_manova.R
\name{print.gwmmr_manova}
\alias{print.gwmmr_manova}
\title{Print Method for GWMMR MANOVA Objects}
\usage{
\method{print}{gwmmr_manova}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"gwmmr_manova"} returned by \code{\link{manova_gwmmr}}.}

\item{...}{Further arguments passed to or from other methods. Currently ignored.}
}
\value{
Invisibly returns \code{NULL}. Called for its side effect of printing model output to the console.
}
\description{
Prints a brief summary of a \code{gwmmr_manova} object,
including the degrees of freedom used in the Pillai trace F-approximation
and the first few rows of the local multivariate test results.
}
\seealso{
\link{manova_gwmmr}
}
