% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gwmmr_model.R
\name{print.gwmmr_model}
\alias{print.gwmmr_model}
\title{Print Method for GWMMR Model Objects}
\usage{
\method{print}{gwmmr_model}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"gwmmr_model"} returned by \code{\link{gwmmr}}.}

\item{...}{Further arguments passed to or from other methods. Currently ignored.}
}
\value{
Invisibly returns \code{NULL}. Called for its side effect of printing model output to the console.
}
\description{
Prints a brief summary of a \code{gwmmr_model} object,
including the model formula, estimated local coefficients, standard errors,
Pillai Trace Statistics, local p-values,
and model fit statistics such as AIC, AICc, and effective degree of freedom.
}
\seealso{
\link{gwmmr}
}
