% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.TSGWML_model.R
\name{print.TSGWML_model}
\alias{print.TSGWML_model}
\title{Print Method for TSGWML Model Objects}
\usage{
\method{print}{TSGWML_model}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"gwglm_model"} returned by \code{\link{gwglm}}.}

\item{...}{Further arguments passed to or from other methods. Currently ignored.}
}
\value{
Invisibly returns \code{NULL}. Called for its side effect of printing model output to the console.
}
\description{
Prints a brief summary of a \code{TSGWML} object, including the function call, model formula,
global and local variables, global coefficient estimates, summary statistics
of local coefficients, and the likelihood ratio test (LRT) result.
}
\seealso{
\link{TSGWML}
}
