% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sub2ind}
\alias{sub2ind}
\title{Convert subscripts to linear indices in column major order}
\usage{
ind <- sub2ind(i,j,n), then A[i,j] == A[ind] if dim(A)[1] == n
}
\arguments{
\item{i}{row index}

\item{j}{column index}

\item{n}{number of rows}
}
\value{
the linear index.
}
\description{
This function is somewhat the equivalent of the Matlab sub2ind.
}
