% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsgaii.R
\name{nsgaii.calculateHyperVolume}
\alias{nsgaii.calculateHyperVolume}
\title{nsgaii.calculateHyperVolume Calculates the hypervolume of a given pareto front
with respect to a reference point}
\usage{
nsgaii.calculateHyperVolume(pareto, RefPnt)
}
\arguments{
\item{pareto}{This is the pareto front}

\item{RefPnt}{This is the reference point. THat is a point htat is dominated by every solution
in the pareto front}
}
\value{

}
\description{
nsgaii.calculateHyperVolume Calculates the hypervolume of a given pareto front
with respect to a reference point
}
